requirements.txt
Twisted>=13.1.0
cyclone>=1.1
fs>=0.4.0
PyNaCl>=0.1.0
